requirements.txt
torch
einops
torchsummary
torchvision
tqdm
av
numpy
tensorboard
PyYAML
yacs